Move TestServer to its own package to work around import cycles
